State/Methods separation (#3873)
* State/Methods separation: user-facing API changes

The only user-facing API change in this PR is the
addition of two fields to `options.proto`:
* the existing `state` annotation gains the
  ability to say which services it `implements`.
* the existing `service` annotation re-gains the
  ability to say which `state` it is linked to.

If the users don't set the fields, semantics will
remain entirely unchanged. If they do set the
fields they'll gain two new powers:
* They'll be able to implement multiple `service`s
  per state `message`.
* They'll be able to name their `service`s
  independently from their `message`s, which also
  unlocks the ability to put them in different
  files and different packages.

* Minor formatting change

This simply unindents a piece of code in
`state_managers.py` without changing anything
else.

* No-op cleanups

This changes no functionality.

* Rename some methods to be more descriptive.

* No-op: `options.py` API cleanup

This makes the `options.py` API simpler to use and
more constent, but doesn't change any
functionality.

* Fixes to Reboot Bazel rules

1. Allow `py_reboot_library` to have dependencies
   outside its own `.proto` file; necessary for a
   state definition in one proto file to be able
   to depend on method definitions in a different
   `.proto` file. To do so, we rename `deps` to
   the more correctly-named `proto_library` (which
   will itself have necessary proto `deps` set),
   and add a new parameter called `py_deps`.

2. Allow `js_reboot_test` to be passed `env`
   variables and a `size`. This will be used to
   build forthcoming tests.

3. Handle `js_reboot_react_library` edge case:
   don't crash when there are no `message`s in a
   `.proto` file. This is important for edge cases
   where a `.proto` file only supplies `service`s,
   for example.

* Small improvements to `py_web_test_suite`

Two changes:
1. Makes it possible to run an against-Envoy web
   test in-process. This is important for our
   forthcoming test suite for `protoc` that relies
   on dynamically-generated classes that can't be
   pickled and therefore can't be cross-process.
2. Make it possible to pass a `bundle.js` that
   doesn't live under `tests/reboot/web`.

* `ServiceName` -> `StateTypeName` when calling

This is a very large, very boring change.

In the next commits, we'll decouple the naming of
state `message`s from the naming of their
`service`s. We thereby lose the ability to assume
that a gRPC call to `FooMethods.Something()` is
directed at a state called `Foo`. This
now-incorrect assumption underpinned a _lot_ of
our code.

This commit reworks every piece of code that needs
to know the _state_ name, not the _service_ name
to correctly be passed that `StateTypeName`
instead of a `ServiceName`.

Most fundamentally this means changing our
`Resolver`s, which unfortunately still means
changing our `PlacementPlanner` and thereby our
`ApplicationConfig`s... Have I mentioned that I
really want to delete all of that code in favor of
only using the routing logic in the router? :D

Beyond those changes this is all very mechanical
and boring.

* Introduce a test suite for `protoc`

This adds all of the tests that we're going to
make pass with forthcoming commits.

We add a whole bunch of new tests in
`tests/reboot/protoc`, in two categories:

 ## Success tests

Each of these is driven by a (set of) `.proto`
files that, in the end, must produce the same
`Echo` API no matter how the `.proto` describes
it.

The `helpers.bzl` then generates three test
targets for each of these: Python, Node.js, and
React.

In all languages, these tests constitute some code
that assumes there to be one or more `Echo`
client(s) in the generated code. These clients are
then exercised to check that they can perform
basic operations.

 ## Failure tests

These take a single `.proto` file that contains
some error. We check that the expected error is
produced. Since all errors are produced by
`protoc_gen_reboot_generic.py` it does not matter
which plugin is actually invoked to produce the
error (we use Python).

* Split Reboot "services" into "states" and "clients"

This is the meat of the work in this PR to split
implementation ("state") from interfaces ("method
definition service").

Roughly, it does two things:

Firstly: in our `protoc` plugin code we were already
generating a bunch of metadata about the services
in our `.proto` files. This commit splits that
metadata into two: information about "clients"
(those parts of the API used to call methods) and "states"
(those parts of the API used to implement
methods). It thereby becomes possible to have
enough information to generate a client without
having access to the state.

Secondly: our generated code templates are updated
to use this `states`/`clients` distinction, and to
generate valid clients-without-states. We also
make sure to use the right package names for each
of the messages/services/methods involved.

With this change it becomes possible to:
1. Have a state `message` in a different file in a
   a differen package
   than its methods `service`.
2. Have a state `message` with multiple methods
   `service`s.

Note that this change is tested by the test suite
that was already introduced in the prior commits.

* Also update boilerplate generation

This allows us to remove the now-obsolete
`BaseFile.services` entry to validate all code
generation now happens either based on a
`BaseClient` or a `BaseState`.

* Update generated Cloud files

* Address review comments