Make Bazel print failed test logs to `stdout` (#3879)
* Make Bazel print failed test logs to `stdout`

A few weeks ago, Bazel used to print the full test
output of failed tests to `stdout`. As of
recently, for some reason I can't find, a failed
test only prints something like...

```
//tests/reboot/controller:placement_planner_test_py                      FAILED in 4.2s
  /home/vscode/.cache/bazel/_bazel_vscode/e9b5149ef40c40531841044f5a70254d/execroot/com_github_reboot_dev_respect/bazel-out/k8-fastbuild/testlogs/tests/reboot/controller/placement_planner_test_py/test.log
```

That makes it impossible to debug a test when it
has failed on GitHub Actions, where we don't have
the contents of `bazel-testlogs/` available for
debugging. That's already pretty annoying in
normal cases (you must run the test on your own
machine before you can start the debugging
journey), but is especially painful in cases of
flakes, which don't reproduce well and where the
error output may contain important hints for how
to reproduce (faster).

This change sets an entry in `.bazelrc` to
reenable the old behavior of printing errors to
`stdout`, and thereby to CI logs.

An alternative approach could have been to collect
and upload `bazel-testlogs/`, but that would still
require the extra step of finding and downloading
the right log file before you can debug a CI
failure.

* Address review comment