Fix several `rbt serve` UX issues (#3917)
* Fix chart UX issue that causes test flakiness

Our `//charts:test_against_kubernetes` test is
very flaky on CI, because the pod it launches as
part of its test will error out within seconds
with an error along the lines of...
```
rbt serve: error: found more than one instance of --directory=VALUE
```

This highlights an actual UX issue with our Helm
chart; the chart isn't usable with applications
built like our `hello` example, because both the
`reboot/examples/hello/.rbtrc` and the chart's
`deployment.yaml` set the `--directory` flag.

The solution in this commit is as brainstormed in
https://github.com/reboot-dev/respect/pull/3914#discussion_r1834865681:
making `rbt serve` the `Dockerfile`'s `ENTRYPOINT`
so that args can be set in e.g. the chart, while
removing such environment-dependent args from the
`.rbtrc`.

Note however that this commit breaks usage of
these `Dockerfile`s on Reboot Cloud and other
platforms that expect to set a `PORT` environment
variable; a follow-on commit in this PR will fix
that.

While we're here: also allow users of the Helm
chart to set the port; that's important if users
ever want to create their own `Service`s, and
reduces the number of hardcoded numbers in the
chart.

* `rbt serve` UX: `--directory`->`--state-directory`

Previously, `rbt serve --directory=foo` was
somewhat ambiguous: what does "serve directory
foo" mean? In most frameworks such wording means
"serve source files stored in `foo`", or "serve
static assets from `foo`", which is not what it
means in Reboot.

This commit renames the `--directory` flag to
`--state-directory` to avoid this confusion, and
edits the few places in the code where we weren't
already using this unambiguous name.

* Rename `rbt serve` -> `rbt serve run`

This matches `rbt dev run`, and leaves us room for
other subcommands like `tasks`, `import`,
`export`, etc. if we want them in the future
(#3426).

* Allow some `rbt serve run` CLI flags from envvars

Before this change, we were in a bit of a bind.

Either we'd set `rbt serve run` flags on the CLI
(e.g. with `docker run [rbt_serve_container]
--port=...` or the equivalent Kubernetes YAML),
but then...
* We _must_ assume that all Cloud containers have
  the `rbt serve run` entrypoint, which some
  currently don't (most notably: our own cloud),
  and...
* Our canonical `Dockerfile` examples don't run on
  e.g. Fly.io or Render.com because those can't
  pass flags to containers.

... OR we'd set flags via `.rbtrc`, but then...
* The `.rbtrc` has to hardcode configuration
  (like the state directory) that should really be
  determined by the operator, with the concomitant
  increased risk of misconfiguration.

This commit adds a better option: it becomes
possible to set some flags from environment
variables, much like e.g. `pip` does. For this
commit those flags are only `--port` (`PORT` or
`RBT_PORT`) and `--state-directory`
(`RBT_STATE_DIRECTORY`), since those are the
operator-determined ports most affected.

So with this change...
1. Reboot Cloud, Fly.io, Render.com, etc. can all
   set their platform-provided `PORT` variable.
2. Reboot Cloud sets the `RBT_STATE_DIRECTORY`
   environment variable, while on other platforms
   the operator can set that envvar manually.