Stop gap for #3908 (#3910)
* Stop gap for #3908

* Support using websockets for reactive readers

This is the first of a phase of commits that will remove the
requirement on TLS for `rbt dev` and thus simplify development.

* Review comments

* Update goldens

* Don't use 'localhost.direct' by default

* Fix tests

1. Use `dev.localhost.direct` in `rbt init` apps
   for now (just like with examples).
2. Correct some `ExternalContext`s running in
   tests.
3. Fix an issue where web tests timed out because
   websocket servers failed to terminate.

* Also cancel query when websocket closes

... Not just on server termination.

* Only wrap serve in a task once waiting for all possible reasons to cancel

* Take a URL instead of a "gateway" when constructing an `ExternalContext`

This was done in a backwards compatible fashion for Python but not for
TypeScript. We should be able to do another pass to remove it for
Python but wanted to keep this change incremental where we have a lot
more existing tests.

* Explicitly check for 'http' or 'https' when constructing Reboot client in React

* Fix another version of the wait-for-client race

In the first commit in this sequence, we
introduced code that waited for the first (and
only) Envoy to connect as a client to the xDS
server, so that we'd be confident that the Envoy
had its config before the test blocks the xDS
server's event loop.

However, we were still observing flakes where
sometimes the Envoy didn't open its listener port.

To understand why, we need to look at the xDS
configuration sequence in more detail:
1. xDS server and Envoy come up
2. Envoy connects to the xDS server
3. Envoy requests cluster configs; is eventually
   sent those.
4. Envoy requests listener configs; is eventually
   sent those.

Before this commit, even though we were waiting
for a first client to be connected and to be fully
up-to-date, we were considering our first client
to be fully up-to-date when it had received all of
the configs it had _requested_; it was therefore
possible that `_wait_until_all_clients_at_version`
returned after (2) already, when the client hadn't
requested (or received) any configuration yet.

This commit fixes that by assuming that all
clients will eventually want both the cluster and
listener configs, and that clients aren't up to
date until both of those have been requested and
acknowledged.

* Mitigation for #3944

By increasing Envoy's buffer size for HTTP1.1
connections to 4MiB we unbreak `__/inspect` for
HTTP1.1 users.

* Update examples to not use 'dev.localhost.direct'

... but use HTTP/1.1 instead.

While we're here, also:
* Update `examples/hello-nodjes`'s `hello.proto`
  to match the one in `examples/hello`; there's no
  reason for them to be different.
* Update `counter`'s documentation to call out
  that there are _two_ frontend options.
* Update `counter/web-vite`, `bank`, and `docubot`
  examples to use a `.env` file like all the other
  examples do.

* Review feedback: recommend using a url with ExternalContext, and fix client provider snippet.

---------

Co-authored-by: Robert-Jan Huijsman <22160949+rjhuijsman@users.noreply.github.com>
Co-authored-by: Stu Hood <stuhood@gmail.com>