Add a servicer that serves a static page for `/` (#4007)
* Make `google.api.http` path `/` possible, for us

Before this commit, using the `/` path in a
`google.api.http` annotation would result in
broken websockets in `rbt dev run` (#3190). This
commit fixes the Envoy misconfiguration that
caused that, clearing the way for us to use this
path to solve #2505.

* Add a servicer that serves a static page for `/`

Fixes #2505 - which was observed (again) in recent
user testing with Felix.

Now, when a user clicks the URL displayed in their
console by `rbt dev run` or `rbt cloud up`, they
will see a friendly (static) web page that
explains that they've reached an API (as opposed
to their frontend).

(Given my limited frontend skills I had ChatGPT
write the static page, and it ignored my requests
to make it look like our existing style sheets. We
can circle back to that later, if it's important.)

* Simplify `rbt dev run` console message

Now that the URL for the API is clickable, we
don't need to hesitate to print that full URL. We
can also drop extra instructions about how to use
the URL in an `ExternalContext`:
* Most users will not need to see that every time
* If they click the URL they'll get the same info
* It's not clear the `ExternalContext` is the most
  important entry point to our API anyway.

* Fix rebase