Add `rbt cloud logs` and all underlying infra (#4167)
* `ConfigMap`s in `kubernets_utils`

This allows the Reboot controller to create
`ConfigMap` objects, which we need to configure
the forthcoming FluentBit pod.

TESTED: by subsequent commits in this PR.

* Two improvements to `kubernets_utils.deployments`

1. Allow a `Deployment` to have multiple
   containers.
2. Allow a `Deployment` to mount more types of
   volumes: not just `DynamicVolumeMount`s but
   also config maps and host paths.

These changes are required for us to deploy the
FluentBit sidecar container for log ingestion in a
subsequent commit.

* Introduce the "facilitator" and run it

This introduces the `Facilitator` API, which (for
now) only exists to collect logs. The
implementation is deferred until a later commit.

This then also sets us up to run such a
facilitator for every Reboot application in the
cloud.

* Add implementation for the facilitator

The facilitator's methods are implemented, most
notably to:
1. Be able to ingest logs.
2. Be able to push logs to a `SortedMap` for
   archival (see comments in the code on why this
   split).
3. Be able to retrieve the logs, regardless of
   whether they are in the archive or not.

* Facilitator authn/authz

This uses Istio's existing capability to tell us
the service account identity of the caller to give
us cross-app authn/authz.

* Ingest logs into the facilitator via FluentBit

This deploys a FluentBit sidecar with every
consensus to gather their logs and send them to
the facilitator's logs ingestion API.

* Add a public logs API

This is the API that CLIs, web interfaces, or
anything else used by our customers will call. It
exists to authenticate the customer before calling
internal APIs.

* Add `rbt cloud logs` CLI command

* Add a logs health check to Pulumi

This validates that all of the infrastructure in
the previous PRs works when deployed to our local
Kubernetes clusters.

* Review: expand FluentBit output

Adds UUID and relative order to log records
produced by FluentBit, so that logs can be
ingested idempotently.

* Review: change way of ingesting and retrieving logs

* Fix mismatch between Pulumi and CLI code

The two need to agree on whether an application's
space ID is named after the user or the
application. The backend doesn't care, but if they
don't agree the CLI can't talk to applications
that Pulumi created.

* Review: move ping logs check out of Pulumi

... into the integration test

* Move auth into its own TokenVerifier and Authorizer

* Add reminder-assertion to stubs.py

* Address review, notably switch to UUIDv7

* Address review comment

* Deflake '//tests/reboot/web/test_auth'

* Deflake 'test_websockets_connection'

* Fix after rebase

* Deflake test_reboot_bank