Remove dependency on eventuals (#4220)
* Replace eventuals::expected with tl::expected

This change should enable us to drop eventuals as
a build dependency. This simplifies the build
process and also should reduce the work that will
be needed for future upgrades to more modern
versions of Bazel.

* Remove eventuals from build system

This requires pulling various dependencies into our
repo that previously were transitively satisfied by
eventuals.

For now, we still preserve the eventuals submodule,
but that will be removed in a subsequent PR.