Remove eventuals submodule (#4251)
This requires adding some of the nested submodules
in eventuals as submodules of the top-level mono
repo, in particular stout and bazel-rules-jemalloc.

Note that the nested pyprotoc-plugin submodule
inside eventuals actually pointed at a different
commit (09f0a63a39485e) than the version of the
pyprotoc-plugin submodule in the top-level mono
repo (b06983bb2af17). That just meant that the
version in eventuals was missing two commits.