Upgrade Bazel rules_webtesting (#4261)
The 0.3.5 release of rules_webtesting has a bug
that prevents using it with Bazel >= 7 [1].
However, the subsequent 0.4.0 release of
rules_webtesting is only offered as a Bazel
module. Therefore, work around this issue by using
a Git commit that fixes the bug but predates the
move to Bazel modules.

For some odd reason, this requires adding
rules_webtesting's "Go internal repositories" in
order to resolve a Bazel error:

  Repository '@com_github_gorilla_mux' is not
  defined and referenced by
  '@io_bazel_rules_webtesting//go/wtl/proxy/driverhub:go_default_library'

Adding some extra repositories is probably
harmless, though...

[1] https://github.com/bazelbuild/rules_webtesting/pull/469