Address Bazel 6 incompat in @node_modules location (#4299)
When using Bazel 6, expressions like

  $(locations //:node_modules/some_module)

behave differently: in Bazel 5, this returns a single path, but in Bazel
6, it returns two paths. One path looks like we'd expect, e.g.,:

  bazel-out/k8-fastbuild/bin/external/com_github_reboot_dev_reboot/node_modules/@bufbuild/protoc-gen-es

But the other path is new:

  bazel-out/k8-fastbuild/bin/external/com_github_reboot_dev_reboot/node_modules/.aspect_rules_js/@bufbuild+protoc-gen-es@1.3.2_at_bufbuild_protobuf_1.3.2/node_modules/@bufbuild/protoc-gen-es

The `.aspect_rules_js` subdirectory still exists in Bazel 5, but it is
located within `bazel-bin` and isn't returned by $(locations). My guess
is that this behavior results from a change in sandbox behavour in Bazel
6.

For most purposes, either path should work (they both contain symlinks
to the same underlying content). So workaround this issue by either
using all of the returned paths (e.g., when adding entries to $PATH and
$NODE_PATH) or only using the first returned path.

Along the way, replace $(locations) with $(execpaths), as the former is
deprecated.