Upgrade to rules_esbuild 0.19 (#4359)
rules_esbuild 0.17 introduced an esbuild plugin
that prevents traversing paths outside the Bazel
sandbox [1]. For reasons that are unclear to me,
this breaks our build -- e.g.,
reboot/inspect/index.tsx attempts to 'import
"./index.css"', which fails. For now, just disable
the plugin, because the hermetic build concerns
that the plugin is supposed to address is not
really a top priority for us at the moment anyway.

Why upgrade rules_esbuild? Aside from the general
advantages of fresher dependencies, rules_esbuild
>= 0.20 uses/requires bazel_lib 2, which is also a
dependency of recent versions of rules_oci. So
we'll need to tackle upgrading past rules_esbuild
0.17 at some point anyway, in order to upgrade
rules_oci.

[1] https://github.com/aspect-build/rules_esbuild/pull/160