Migrate from rules_docker to rules_oci (#4353)
We previously used rules_docker to build container
images, but rules_docker has been unmaintained for
several years. Morever, rules_docker has
incompatibilities with Bazel >= 6 and other
dependencies that have prevented us from some
critical upgrades (e.g., gRPC).

This commit replaces rules_docker with rules_oci.
Because rules_oci is not a 1-to-1 replacement for
rules_docker, we need to introduce some extra
machinery:

(1) `dockerfile_oci_image` is a replacement for
    `dockerfile_image`. To use docker to build an
    OCI tar from a Dockerfile, the (non-default)
    `docker-container` builder must be used, so we
    create that at build-time if necessary.

(2) `py_oci_image` is a replacement for
    `py3_image`.

(3) To replace `container_test` in `rules_docker`,
    we use Container Structure Tests, which
    provides a similar `container_test` rule.

We currently use rules_oci 1.8.0, because
rules_oci >= 2.0.0 does not support Bazel 5.
Upgrading to a more recent release of rules_oci
will be done after we move to Bazel 6.

This commit downgrades rules_pkg to 1.0.1: for
some reason, we now hit a codepath in rules_pkg
1.1.0 that is incompatible with Bazel < 7.

We also need to add faux dependencies on gazell
and rules_go, because apparently one of our
dependencies uses them but does not properly
declare its own dependency. That will be addressed
in the future, either as part of upgrading
dependencies and/or moving to Bazel modules.