Bazel: allow fallback to `local` sandbox strategy (#4385)
This is necessary with Bazel 7, because the logic
for having certain rules auto-fallback to local
sandboxing has been removed. We should ideally
look at places where we need local sandboxing
and improve them -- I took a quick look at that
but it does not seem trivial, so I will defer
that for now.

This change is necessary to build with Bazel 7,
but should be harmless (and silence a build
warning) in prior Bazel versions.