Config pods contribute logs to `rbt cloud logs` (#4358)
* Make fluentbit code reusable (code movement only)

This doesn't make any functional changes, it just
moves code out into a file where it can be reused
by other parts of the code.

* Config pods contribute logs to `rbt cloud logs`

Before this commit, config pods' logs were not
captured by `rbt cloud logs` yet, which means that
a fatal application error (e.g. an `ImportError`)
at config-pod-time would result in `rbt cloud up`
failing without a way to diagnose the issue.

With this commit config pods now contribute logs
to `rbt cloud logs`. The primary changes include:
1. Creating config pods as a `Deployment` rather
   than a `Pod`, so we'll be able to reuse the
   code that sets up log capturing for
   consensuses (which are `Deployment`s); this
   doesn't produce any meaningful behavioral
   change.
1. Setting the Kubernetes label that causes the
   Reboot routing filter to activate (so that logs
   can be sent to the facilitator application); we
   rename this label to match its new broader use.
1. Adding a `fluentbit` container to the config
   pod deployment, just like we already did with
   consensuses.

DESIGN CONSIDERATION: An application's config pod
may come up at the same time as its
_facilitator_'s config pod. This means that when
the config pod fails it may not immediately be
able to flush its logs. We believe that is OK,
because:
1. The `fluentbit` sidecar will retry flushing the
   logs until its pod is terminated, which only
   happens after a generous (30-second) timeout.
2. The facilitator's config pod will definitely be
   successful (it does not contain customer code)
   and the facilitator will therefore be very
   likely to be available to receive the logs
   before the application's config pod is
   terminated.
3. The facilitator will in fact be likely to be
   running even before the application's config
   pod is _started_: the facilitator is a
   frequently used image that is almost always
   already cached on the node, while the
   application's container image must still be
   pulled from the registry.

We therefore threw additional measures off the
truck; if we find that logs are not being captured
we can in the future additionally...
* Add a mechanism in the controller that waits
  until logs are flushed before terminating the
  config pod.
* Add a mechanism in the controller that ensures
  the facilitator is running before starting the
  application's config pod.
* [... TBD, potentially more options ...]

* Small cleanups in Fluent Bit code

1. Consistently call it "Fluent Bit" and
   "fluent-bit", like the maintainers do.
2. Turn a few repeated literals into constants.

* Make Pulumi health checks time out eventually

Currently, the gRPC health checks used to verify
e.g. our Cloud app will wait forever for the
target app to become healthy. This isn't very
useful, since:
1. It doesn't give a clear signal when there is an
   issue that's causing the Cloud (or other Reboot
   app) to not start up properly.
2. It makes `Ctrl+C` the only way out, which
   leaves the deployment in an uncertain state
   (requiring the operator to run `[target]
   cancel` before anything else can happen).

This commit adds a simple timeout, so that the
operator is clearly told that something is amiss
when the application fails to start in a timely
manner, and future `[target] up` commands can be
run without a `cancel` in between.

* Make gRPC health checks JSON-transcodable

Before this commit, the only way to do a gRPC
health check was with a gRPC client. This commit
makes it possible to do a health check with a
regular HTTP/JSON client. We'll use this
capability in the next commit.

* Make routing filter return 503 if app is not found

... to ensure that this is considered a retryable
error, not a permanent failure. Apps can come up
at any time, and we don't want to consider them
permanently unavailable if they're not currently
up.

* Pull Fluent Bit container from our own registry

... instead of reaching out to `docker.io` at
prod-time, which is sensitive to rate limits
imposed by `docker.io`.

We achieve this by pulling the image from
`docker.io` at build-time using Bazel and pushing
it to our own cluster's registry just like we do
with other Bazel-built containers.

* `wait_for_running` -> `wait_for_ready`

In all use cases where we said that we want to
wait for a `Pod` to be "running" what we
_actually_ want is to wait for the `Pod` to reach
the "Ready" state - which is not the same - a
`Pod` can be "running" but not have passed its
health checks yet. This caused issues when, during
the development of this chain of commits, we
experimented with using health checks.

This commit therefore updates the
`wait_for_running` method to `wait_for_ready`.

* Fix edge cases in config pod log ingestion

Specifically:
1. Switch to using Istio's "native sidecar"s,
   which ensures that the proxy is up and running
   (so won't produce spurious 404s) _before_ any
   other pods are started.
1. Also make `fluent-bit` a Kubernetes native
   sidecar, to ensure that it is terminated after
   the main application's container, but before
   Istio's proxy, so all logs have a chance to get
   flushed.

This commit also adds a unit test to prove that
the config pod logs ingestion indeed works.
Running it repeatedly shows it works _reliably_
(no flakes).

* Fix out-of-order Bazel load

* Address review comments