Remove outDir from some tsconfig files (#4389)
This is ignored when the project is built using
Bazel and `ts_project`. More recent versions of
`rules_ts` complain if `outDir` is specified in
`tsconfig.json` but not also duplicated as a
parameter to `ts_project`. Since "outDir" was
being ignored anyway, seems simpler to just remove
it.

The one exception is in `reboot/nodejs`, where
"outDir" is *not* being ignored. The root cause
appears to be that this package is not built like
all the others, because of the dependency on gyp.