Improve $(execpaths) robustness in Bazel rules (#4485)
We previously used expressions like $(execpath //some-npm-package) to
find the path where rules_ts has placed an npm package in the execroot.
This doesn't seem to be very robust:

  * It works as we expect with Bazel 5 and rules_ts 2.3.0

  * With Bazel 6 and rules_ts 2.3.0, it returns multiple paths
    (e.g., one referencing a `.aspect_rules_js` subdirectory and the
    other the path we get in Bazel 5). We previously adapted to this
    in 1e5786dd6831088 by switching to $(execpaths).

  * With Bazel 6 and rules_ts 3.5.1, it returns *bizarre* paths in some
    cases: e.g., "$(execpaths //:node_modules/@reboot-dev/reboot-api"
    returns

      [...]/node_modules/.aspect_rules_js/@reboot-dev+reboot-api@0.0.0/node_modules/@bufbuild/protobuf
      [...]/node_modules/.aspect_rules_js/@reboot-dev+reboot-api@0.0.0/node_modules/@reboot-dev/reboot-api
      [...]/node_modules/@reboot-dev/reboot-api

I don't fully understand why this behavior occurs. However, it seems
that `npm_link_package` generates a `.../dir` target that identifies the
directory where the package can be found in a way that seems more
predictable than using $(execpath) on the package name itself.