Move part of `reboot/aio` to the `public/` folder (#4531)
* File reorganizations

In preparation for making much of `aio/` public,
move some files around.

1. Move `controller/exceptions.py` to `aio/`; in
   practice it was used there, and its exception
   is not limited to the controller.
2. split `aio/external.py` into two files: the
   part with the `ExternalContext` is to become
   public (stays where it is), while a helper
   method that shouldn't become public is moved to
   the `//tests` package. This is to avoid
   `//kubernetes_utils` (a dependency of the
   helper) having to become public.
3. Move `chdir()` to its own file. This little
   function (without deps) is the reason
   that `reboot/aio` depends on `reboot/cli`. By
   moving it into its own file we break that
   dependency.

* Stop using 'cli/terminal.py' in protoc plugin

It was already questionable to use
`cli/terminal.py` from protoc plugin code, as it
seems to assume that the protoc plugins will
always be run from the `rbt` CLI, which is untrue.
However, it was also the last remaining place
where non-CLI code depended on CLI code, so
breaking this dependency (by replacing the code
with a simple `print()` and `exit()`) makes it
much easier to open source only limited parts of
our codebase at a time.

Also stop using terminal in the other protoc plugins

* Rebased: moving large parts of aio

* Also move the new 'ssl/' directory

* Add a missing package to our Codespaces

* Fix a test issue caused by gRPC version mismatches

* Fixes to make type checking of generated code work again, post-move

* Add more untyped Google libraries to our example's '.mypy.ini' ignore list

* Fix input_error_integration_test

* Update goldens

* Fix external_context_tests