Allow Cloud customers to use their API keys for admin auth (#4533)
* Always generate Python `_rbt.py` files

... Even if there are no Reboot services in the
proto file being compiled. This is important when
a developer wants to use a _legacy_ gRPC service
with a Reboot application, but there are no Reboot
services in the proto file containing that legacy
gRPC service - we need the `_rbt.py` file for
Reboot to be able to use the legacy gRPC service.

This behavior is important for the `AdminAuth`
servicer that will be introduced in a subsequent
commit - it will be a legacy gRPC service in a
proto file that does not contain any Reboot
servicers, yet it will run as part of a Reboot
application.

* Also print stack traces when _legacy_ gRPC calls fail

* Move 'forwarded_client_cert_auth.py'

... to a shared location, where it can be used by
both `cloud/facilitator/` code and other
(upcoming) `cloud/` code.

* Implement API key auth for admin endpoints

The code in this commit works as follows:
1. A request comes in to `admin_auth.py` in the
   customer's application as it did before.
2. If a new `REBOOT_ADMIN_AUTH_APPLICATION_ID`
   environment variable is set, `admin_auth.py`
   will call the new `AdminAuth` gRPC service in
   the specified application to ask whether the
   given bearer token identifies one of its
   admins.
3. The application running that `AdminAuth`
   service will be our "cloud" application.
4. It will use Istio's SPIFFE header to determine
   the calling application's ID, and uses its
   access to the `APIKey` and `Application` states
   to determine whether the given bearer token is
   a valid admin API key for that application.

* UI: rename "Admin Bearer Token" and "Admin Secret"

... To "Admin Credentials", in CLI and web UI. The
names "bearer token" (too implementation-specific)
and "admin secret" (not always correct) are not
ideal; when using an API key, neither of these
terms feel "right". The term "credentials" is more
generic, and someone holding either an API key or
any other pre-shared secret will naturally
understand that's what's meant by "credentials".

* Mention inspect endpoint on the root page

This is a page that customers are likely to click
right after `rbt cloud up`, as it prints its URL.
Now that the inspect page works on Cloud, we can
mention it on that page again.

Fixes #4460.

* Fix auth docs title

In the `docs.reboot.dev` sidebar, the `auth.mdx`
title is shown as "auth" instead of the desired
"Authentication". That's because the first
markdown statement in the file is a comment, not a
header. Moving the comment below the header fixes
the title in the sidebar.

* Address review feedback