Add tasks to the inspect dashboard (#4751)
* Bugfix: return complete error message

Before this fix, the error message returned by
`_unknown_query_or_mutation_error_message` was in
fact only the first line of the fully written
error message (notice the `,` at the end of the
last line...)

* API fix: make `reboot.aio.tasks` importable

In moving to `rebootdev.aio.tasks` we'd failed to
leave a `reboot.aio.tasks` forwarder in place,
which means the public API inadvertently changed.
This hadn't been noticed, because the
`reboot.aio.tasks` API isn't used in any of our
public examples.

* Add a task to the `ping` application

To iterate on the visual design of a tasks
dashboard, one needs to have a running Reboot
application that's running one or more tasks.
Previously, we didn't have such an application
easily available (e.g. none of our examples use
tasks).

The `ping` application is a good candidate to be
the app to use during development, because it is
bazilified. This commit...
1. Makes `ping` runnable using `ibazel`, allowing
   for a tight dev loop.
2. Adds a task to it, which can be used to
   manually test the tasks dashboard.
3. Adds a unit test, to ensure the app works.

This addition doesn't intefere with existing uses
of the `ping` application (e.g. in `cluster_test`)
because the task runs on a different state.

* Add a new `ListTasksStream` RPC

This will be what the dashboard will call to watch
the list of tasks as it changes.

Makes the RPC callable via HTTP/JSON by adding a
`TasksRoutable` to `servers.py`.

Also adds a `SCHEDULED_ITERATION` state to tasks,
since it's unhelpful to say that a loop that's
waiting to run its next iteration (possibly doing
nothing for many hours) is `STARTED`; it's more
useful to say that it's scheduled to run its next
iteration.

* Add tasks to the inspect dashboard

This expands the "inspect" dashboard to show not
only states, but also recent tasks. Its display
closely follows the layout of the `rbt tasks`
CLI.

See screenshots in the PR description.

* Rebase on latest workflow changes, and fix #4761

Use the latest workflow looping API, and fix #4761
where the last iteration of a workflow was not
being counted.

TESTED: new unit test in `tasks_servicer_tests.py`
        would have failed previously, now passes.

* Address review comments

* update goldens