Custom propagation of tracing information
Before this change, we would use OpenTelemetry's
gRPC interceptors to propagate tracing
information. This led to a number of issues:
* We couldn't suppress tracing for gRPC calls,
  tracing context propagation and RPC span
  creation were a package deal.
* There are bugs in the interceptors that cause
  log spam, while other bugs have previously
  required us to build workarounds in
  `tracing.py`.

This change replaces the use of OpenTelemetry's
gRPC interceptors at the `CUSTOMER` tracing level
(the default) with propagation of tracing
context via our `headers.py`.

This commit does NOT create spans for gRPC calls,
which makes the dashboard less noisy and more
focused on customer code.

At the `REBOOT_INTERNAL` tracing level we still
use the OpenTelemetry gRPC interceptors, since we
want to see every gRPC call in that case.