Minimal fix for #4630 (`rbt cloud down` did not expunge) (#4813)
* Minimal fix for #4630

In #4630, `rbt cloud down` does not expunge the
data of Cloud applications, even though it should.
This commit makes it so that it does again.

* Have `local`` storage work like EFS

Before this commit, the local storage used by our
dev clusters was per-consensus, unlike the EFS
storage which was per-application.

Since dev clusters only have one node, it's OK to
have the local storage also work per-application
even though it's not a distributed storage. By
having the local setup echo that of EFS (which we
use in prod) we're much more likely to catch bugs
related to `PersistentVolumeClaim` usage early,
ideally even in our tests.

Using this more-symmetric setup, we can use a dev
cluster to confirm that the fix in the previous
commit works: expunging now works on the Cloud.

* Add websockets option to the reboot-web

* Don't backoff if aborted

* Prepare 0.33.1 for release

---------

Co-authored-by: Zakhar <zpilipuk@gmail.com>
Co-authored-by: Benjamin Hindman <benjamin.hindman@gmail.com>