`rbt cloud logs`: default to `--revisions=">=latest"` (#4830)
* Backwards compatibility: allow `(u)int32`->`(u)int64`

In protobuf it's safe to convert `int32` and
`uint32` fields to `int64` and `uint64`
respectively, but our backwards compatibility
policy didn't allow for that.

This commit relaxes our checks so that this type
of change is allowed. This will permit a future
cleanup between `application.proto` and
`logs.proto` where there is a mismatch in how
`revision_number` is typed; one uses `uint64`
while the other uses `uint32`.

* Facilitator: small bugfix in `LogsServicer` cleanup

Before this commit, the `LogsServicer` would not
properly cancel all tasks when a running
`--follow` RPC was cancelled, resulting in some
log spam in its Kubernetes pod logs. This commit
moves the appropriate cleanup logic into a
`finally` block to ensure that it always runs.

* 'rbt cloud logs': default to '--revision=">=latest"'

Per user feedback from Moment.

* Rename variables